requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
APScheduler==3.10.4
asyncio==3.4.3
attrs==24.2.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gensim==4.3.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
nltk==3.9.1
numpy==1.26.4
openai==1.51.0
packaging==24.1
pillow==10.4.0
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.1
scipy==1.13.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
telegram==0.0.1
textblob==0.18.0.post0
thinc==8.3.2
tokenizers==0.20.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.31.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.13.1
